Optimizacion de código para que no se exceda del 100%

+++ Dispensador/lcd_ui.h
@@ -2,5 +2,5 @@
 
 #include "config.h"
 
-void setLcdState(LcdMode mode, int compartimento, int cantidad, const char* medicamento);
+void setLcdState(LcdMode mode, int8_t compartimento, uint8_t cantidad, const char* medicamento);
 void updateLCD();
